Install from code artifact (#12)
## Why is this change necessary?
Need a configuration option to be able to install from CodeArtifact


 ## How does this change address the issue?
Adds a step in CI to authenticate with the OIDC role. Creates scripts to
be able to authenticate locally. Updates pyproject.toml to specify
CodeArtifact


 ## What side effects does this change have?
None


 ## How is this change tested?
In downstream repos that use CodeArtifact and those that don't

For CI in this repo, it was deemed prohibitive to try and actively
install from CodeArtifact, so there's a script that updates any
instantiated templates pyproject.toml to point it back to PyPI